requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
djangorestframework-filters==0.7.0
django-environ==0.8.1
django-extensions==3.2.1
pytest==7.2.2
pytest-django==4.5.2